Refactor color and size modifiersto another module
