+++ geoguide/client/templates/geoguide/environment.html
@@ -16,7 +16,7 @@
   <div class="controlContainer" id="customControlsLeftTop">
     <div class="control well well-sm">
       <div class="title" data-toggle="collapse" data-target="#collapseGeoGuideParameters">GeoGuide Parameters <span class="pull-right"><span class="caret"></span></span></div>
-      <div class="collapse in" id="collapseGeoGuideParameters">
+      <div class="collapse" id="collapseGeoGuideParameters">
         <div class="form-horizontal">
           <div class="form-group">
             <label for="colorModifier" class="col-sm-4 control-label">Color modifier:</label>
@@ -49,7 +49,7 @@
     </div>
     <div class="control well well-sm">
       <div class="title" data-toggle="collapse" data-target="#collapseIUGADefinition">Highlighting Parameters <span class="pull-right"><span class="caret"></span></span></div>
-      <div class="collapse in" id="collapseIUGADefinition">
+      <div class="collapse" id="collapseIUGADefinition">
         <div class="form-horizontal">
           <div class="form-group">
             <label for="timelimit" class="col-sm-4 control-label">Max delay (ms):</label>
